requirements.txt
outlines
llguidance
llama-cpp-python
Flask 
flask_cors
requests
sseclient-py
# pip install "smolagents[toolkit]"